Fix signiture of old syscalls

+++ src/nautilus/syscalls/open.c
@@ -1,10 +1,10 @@
 #include <nautilus/nautilus.h>
 #include <nautilus/fs.h>
 
-int
-sys_open(int filename,
-	 int     flags,
-	 int  mode)
+uint64_t
+sys_open(uint64_t filename,
+	 uint64_t     flags,
+	 uint64_t  mode)
 {
 	struct nk_fs_open_file_state *file;
 	file = nk_fs_open((char*)filename,flags, 0);
